- Fixed avatar shop

+++ MajesticLife/app/src/main/java/com/farazannajmi/majesticlife/DataStructures/MajesticLifeRoomDatabase.java
@@ -9,6 +9,7 @@ import android.os.AsyncTask;
 import android.support.annotation.NonNull;
 import android.util.Log;
 
+import com.farazannajmi.majesticlife.DataHolder;
 import com.farazannajmi.majesticlife.LoadingActivity;
 import com.farazannajmi.majesticlife.R;
 
@@ -20,10 +21,10 @@ import com.farazannajmi.majesticlife.R;
  * https://medium.com/@srinuraop/database-create-and-open-callbacks-in-room-7ca98c3286ab
  */
 
-//@Database(entities = {Word.class}, version = 1)
 @Database(entities = {User.class, Faaliat.class, Skill.class, Quest.class, PlanCell.class,
-        FaaliatSkill.class, QuestSkill.class, FaaliatRepetitions.class},
-        version = 4)
+        FaaliatSkill.class, QuestSkill.class, FaaliatRepetitions.class,
+        Avatar.class, AvatarItem.class},
+        version = 5)
 public abstract class MajesticLifeRoomDatabase extends RoomDatabase
 {
     private static final String DB_NAME = "MajesticLifeDatabase.db";
@@ -118,93 +119,99 @@ public abstract class MajesticLifeRoomDatabase extends RoomDatabase
         @Override
         protected Void doInBackground(final Void... params)
         {
-            if(LoadingActivity.isFirstTime)
+            if(LoadingActivity.isFirstTime && !DataHolder.IsDatabaseCreated)
             {
+                //order is important!!
+
                 //User:
                 userDao.deleteAll();
                 User user = new User(0, "NewKing", "king@mail.com",
                         R.drawable.ic_king, 0, 1,0, 1, 0, 1, 100);
                 userDao.insert(user);
 
+                //Faaliats:
                 faaliatDao.deleteAll();
                 Faaliat faaliat = new Faaliat(0, "Reading book",
                         R.drawable.ic_majestic_activities, R.color.faaliatsColor1, 10,0,1,0);
                 faaliatDao.insert(faaliat);
 
+                //Skills:
                 skillDao.deleteAll();
                 Skill skill = new Skill(0, "Knowledge", R.drawable.ic_skills, 1, 0, 0);
                 skillDao.insert(skill);
 
+                //FaaliatSkills:
                 faaliatSkillDao.deleteAll();
                 FaaliatSkill faaliatSkill = new FaaliatSkill(0, 0, 5);
                 faaliatSkillDao.insert(faaliatSkill);
 
+                //PlanCells:
                 planCellDao.deleteAll();
                 PlanCell planCell = new PlanCell(0, 0, 0,5);
                 planCellDao.insert(planCell);
 
+                //Quests:
                 questDao.deleteAll();
-                Quest quest = new Quest(0, "Pass exam", "20181203"); //todo: wrong DuaDate param
+                Quest quest = new Quest(0, "Pass exam", "2019/1/1"); //todo: wrong DuaDate param
                 questDao.insert(quest);
 
+                //QuestSkills:
                 questSkillDao.deleteAll();
                 QuestSkill questSkill = new QuestSkill(0, 0, 2);
                 questSkillDao.insert(questSkill);
 
-                avatarDao.deleteAll();
-                Avatar avatar = new Avatar(0,0, true, R.drawable.ava_back1,
-                        R.drawable.ava_skin_white, R.drawable.ava_cloth1,
-                        R.drawable.ava_eyes1, R.drawable.ava_mouth1, R.drawable.ava_crown1);
-                avatarDao.insert(avatar);
-
-                //region ---------- initial avatarItems ----------
+                //region ---------- avatarItems ----------
                 //backgrounds:
                 avatarItemDao.deleteAll();
-                AvatarItem ava_back1 = new AvatarItem(R.drawable.ava_back1, true, 0);
+                AvatarItem ava_back1 = new AvatarItem(R.drawable.ava_back1, true, 0, 0);
                 avatarItemDao.insert(ava_back1);
-                AvatarItem ava_back2 = new AvatarItem(R.drawable.ava_back2, false, 0);
+                AvatarItem ava_back2 = new AvatarItem(R.drawable.ava_back2, false, 0, 50);
                 avatarItemDao.insert(ava_back2);
-                AvatarItem ava_back3 = new AvatarItem(R.drawable.ava_back3, false, 0);
+                AvatarItem ava_back3 = new AvatarItem(R.drawable.ava_back3, false, 0, 60);
                 avatarItemDao.insert(ava_back3);
-                AvatarItem ava_back4 = new AvatarItem(R.drawable.ava_back4, false, 0);
+                AvatarItem ava_back4 = new AvatarItem(R.drawable.ava_back4, false, 0, 60);
                 avatarItemDao.insert(ava_back4);
-                AvatarItem ava_back5 = new AvatarItem(R.drawable.ava_back5, false, 0);
+                AvatarItem ava_back5 = new AvatarItem(R.drawable.ava_back5, false, 0, 70);
                 avatarItemDao.insert(ava_back5);
-                AvatarItem ava_back6 = new AvatarItem(R.drawable.ava_back6, false, 0);
+                AvatarItem ava_back6 = new AvatarItem(R.drawable.ava_back6, false, 0, 70);
                 avatarItemDao.insert(ava_back6);
-                AvatarItem ava_back7 = new AvatarItem(R.drawable.ava_back7, false, 0);
+                AvatarItem ava_back7 = new AvatarItem(R.drawable.ava_back7, false, 0, 70);
                 avatarItemDao.insert(ava_back7);
 
+                //Skins
+                AvatarItem ava_skin1 = new AvatarItem(R.drawable.ava_skin_white, true, 0, 0);
+                avatarItemDao.insert(ava_skin1);
+                //todo: insert other clothes
+
                 //Clothes
-                AvatarItem ava_cloth1 = new AvatarItem(R.drawable.ava_cloth1, true, 0);
+                AvatarItem ava_cloth1 = new AvatarItem(R.drawable.ava_cloth1, true, 0, 0);
                 avatarItemDao.insert(ava_cloth1);
                 //todo: insert other clothes
 
                 //Eyes
-                AvatarItem ava_eyes1 = new AvatarItem(R.drawable.ava_eyes1, true, 0);
+                AvatarItem ava_eyes1 = new AvatarItem(R.drawable.ava_eyes1, true, 0, 0);
                 avatarItemDao.insert(ava_eyes1);
                 //todo: insert other eyes
 
                 //Mouths
-                AvatarItem ava_mouth1 = new AvatarItem(R.drawable.ava_mouth1, true, 0);
+                AvatarItem ava_mouth1 = new AvatarItem(R.drawable.ava_mouth1, true, 0, 0);
                 avatarItemDao.insert(ava_mouth1);
                 //todo: insert other mouths
 
                 //Crowns
-                AvatarItem ava_crown1 = new AvatarItem(R.drawable.ava_crown1, true, 0);
+                AvatarItem ava_crown1 = new AvatarItem(R.drawable.ava_crown1, true, 0, 0);
                 avatarItemDao.insert(ava_crown1);
                 //todo: insert other crowns
                 //endregion ------------------------------
 
                 //no initial FaaliatRepetitions
-
                 //todo test:
                 faaliatRepetitionsDao.deleteAll();
                 FaaliatRepetitions fr1 = new FaaliatRepetitions(0, 1, "2018/6/10", 1);
                 FaaliatRepetitions fr2 = new FaaliatRepetitions(0, 2, "2018/6/11", 3);
                 FaaliatRepetitions fr3 = new FaaliatRepetitions(0, 3, "2018/6/12", 2);
                 FaaliatRepetitions fr4 = new FaaliatRepetitions(0, 4, "2018/6/13", 5);
-                FaaliatRepetitions fr5 = new FaaliatRepetitions(0, 6, "2018/6/15",3);
+                FaaliatRepetitions fr5 = new FaaliatRepetitions(0, 6, "2018/6/15", 3);
                 FaaliatRepetitions fr6 = new FaaliatRepetitions(0, 1, "2018/6/17", 1);
                 FaaliatRepetitions fr7 = new FaaliatRepetitions(0, 3, "2018/6/19", 1);
                 FaaliatRepetitions fr8 = new FaaliatRepetitions(0, 4, "2018/6/20", 2);
@@ -217,6 +224,14 @@ public abstract class MajesticLifeRoomDatabase extends RoomDatabase
                 faaliatRepetitionsDao.insert(fr7);
                 faaliatRepetitionsDao.insert(fr8);
 
+                //Avatar:
+                avatarDao.deleteAll();
+                Avatar avatar = new Avatar(0,0, true, R.drawable.ava_back1,
+                        R.drawable.ava_skin_white, R.drawable.ava_cloth1,
+                        R.drawable.ava_eyes1, R.drawable.ava_mouth1, R.drawable.ava_crown1);
+                avatarDao.insert(avatar);
+
+                DataHolder.IsDatabaseCreated = true;
                 Log.d("Data", "Database initialed!");
             }
             return null;
